Refactor: creation of common folder for reusable components; moved Tabs into common folder and updated imports accordingly
